// STD7_SCI/CH3/Ch3_3.2/src/interfaces/circuitTypes.ts
export type Language = 'en' | 'hi' | 'gu';